Made some changes sidd suggested in order to try and get the sphere working. (still broken)

+++ PS4_ClearScreen_JL/PS4_ClearScreen_JL/Model.h
@@ -62,10 +62,10 @@ public:
 	PsShader *psShader;
 
 	std::vector<Vertex> vertices;
-	std::vector<uint16_t> indices;	
+	std::vector<uint32_t> indices;
 	
 	void *fsMem;
-	uint16_t *indexData;
+	uint32_t *indexData;
 	uint32_t shaderModifier;
 
 	sce::Gnmx::InputOffsetsCache vsInputOffsetsCache;
