+++ PS4_ClearScreen_JL/PS4_ClearScreen_JL/Utils.h
@@ -31,10 +31,10 @@ public:
 	Utils();
 	~Utils();
 
-	static void setTriData(std::vector<Vertex>& vertices, std::vector<uint16_t>& indices);
-	static void setQuadData(std::vector<Vertex>& vertices, std::vector<uint16_t>& indices);
-	static void setCubeData(std::vector<Vertex>& vertices, std::vector<uint16_t>& indices);
-	static void setSphereData(std::vector<Vertex>& vertices, std::vector<uint16_t>& indices);
+	static void setTriData(std::vector<Vertex>& vertices, std::vector<uint32_t>& indices);
+	static void setQuadData(std::vector<Vertex>& vertices, std::vector<uint32_t>& indices);
+	static void setCubeData(std::vector<Vertex>& vertices, std::vector<uint32_t>& indices);
+	static void setSphereData(std::vector<Vertex>& vertices, std::vector<uint32_t>& indices);
 
 	
 };
